+++ src/lib.rs
@@ -15,8 +15,6 @@ use build_manager::BuildManager;
 use build_manager::NewJob;
 pub use color_eyre;
 use color_eyre::eyre::eyre;
-#[cfg(feature = "rustc")]
-use color_eyre::eyre::Context as _;
 pub use color_eyre::eyre::Result;
 pub use core::run_and_collect;
 pub use core::CrateType;
@@ -81,10 +79,8 @@ pub fn run_tests(mut config: Config) -> Result<()> {
     #[cfg(feature = "gha")]
     let name = display(&config.root_dir);
 
-    let text = match args.format {
-        Format::Terse => status_emitter::Text::quiet(),
-        Format::Pretty => status_emitter::Text::verbose(),
-    };
+    let emitter: Box<dyn StatusEmitter> = args.format.into();
+
     config.with_args(&args);
 
     run_tests_generic(
@@ -92,9 +88,9 @@ pub fn run_tests(mut config: Config) -> Result<()> {
         default_file_filter,
         default_per_file_config,
         (
-            text,
+            emitter,
             #[cfg(feature = "gha")]
-            status_emitter::Gha::<true> { name },
+            status_emitter::Gha { name, group: true },
         ),
     )
 }
@@ -153,8 +149,7 @@ pub fn default_per_file_config(config: &mut Config, file_contents: &Spanned<Vec<
 pub fn test_command(mut config: Config, path: &Path) -> Result<Command> {
     config.fill_host_and_target()?;
 
-    let content = Spanned::read_from_file(path)
-        .wrap_err_with(|| format!("failed to read {}", display(path)))?;
+    let content = Spanned::read_from_file(path).transpose()?;
     let comments = Comments::parse(content.as_ref(), &config)
         .map_err(|errors| color_eyre::eyre::eyre!("{errors:#?}"))?;
     let config = TestConfig {
@@ -183,7 +178,7 @@ pub fn run_tests_generic(
     mut configs: Vec<Config>,
     file_filter: impl Fn(&Path, &Config) -> Option<bool> + Sync,
     per_file_config: impl Copy + Fn(&mut Config, &Spanned<Vec<u8>>) + Send + Sync + 'static,
-    status_emitter: impl StatusEmitter + Send,
+    status_emitter: impl StatusEmitter,
 ) -> Result<()> {
     if nextest::emulate(&mut configs) {
         return Ok(());
@@ -244,7 +239,8 @@ pub fn run_tests_generic(
                         // Forward .rs files to the test workers.
                         submit
                             .send(Box::new(move |finished_files_sender: &Sender<TestRun>| {
-                                let file_contents = Spanned::read_from_file(&path).unwrap();
+                                let file_contents =
+                                    Spanned::read_from_file(&path).transpose().unwrap();
                                 let mut config = build_manager.config().clone();
                                 let abort_check = config.abort_check.clone();
                                 per_file_config(&mut config, &file_contents);