+++ src/status_emitter/gha.rs
@@ -12,13 +12,10 @@ use std::{
 
 use super::{RevisionStyle, StatusEmitter, Summary, TestStatus};
 fn gha_error(error: &Error, test_path: &str, revision: &str) {
-    let file = Spanned::read_from_file(test_path).unwrap();
+    let file = Spanned::read_from_file(test_path).transpose().unwrap();
     let line = |span: &Span| {
-        let line = file
-            .lines()
-            .position(|line| line.span.bytes.contains(&span.bytes.start))
-            .unwrap();
-        NonZeroUsize::new(line + 1).unwrap()
+        let line = file[..=span.bytes.start].lines().count();
+        NonZeroUsize::new(line).unwrap_or(NonZeroUsize::MIN)
     };
     match error {
         Error::ExitStatus {
@@ -187,19 +184,22 @@ fn gha_error(error: &Error, test_path: &str, revision: &str) {
 }
 
 /// Emits Github Actions Workspace commands to show the failures directly in the github diff view.
-/// If the const generic `GROUP` boolean is `true`, also emit `::group` commands.
-pub struct Gha<const GROUP: bool> {
+/// If the [`Self::group`] boolean is `true`, also emit `::group` commands.
+pub struct Gha {
     /// Show a specific name for the final summary.
     pub name: String,
+    /// Use github grouping/folding feature to collapse individual test outputs and entire test suites.
+    pub group: bool,
 }
 
 #[derive(Clone)]
-struct PathAndRev<const GROUP: bool> {
+struct PathAndRev {
     path: PathBuf,
     revision: String,
+    group: bool,
 }
 
-impl<const GROUP: bool> TestStatus for PathAndRev<GROUP> {
+impl TestStatus for PathAndRev {
     fn path(&self) -> &Path {
         &self.path
     }
@@ -208,6 +208,7 @@ impl<const GROUP: bool> TestStatus for PathAndRev<GROUP> {
         Box::new(Self {
             path: self.path.clone(),
             revision: revision.to_owned(),
+            group: self.group,
         })
     }
 
@@ -215,11 +216,12 @@ impl<const GROUP: bool> TestStatus for PathAndRev<GROUP> {
         Box::new(Self {
             path: path.to_path_buf(),
             revision: self.revision.clone(),
+            group: self.group,
         })
     }
 
     fn failed_test(&self, _cmd: &str, _stderr: &[u8], _stdout: &[u8]) -> Box<dyn Debug> {
-        if GROUP {
+        if self.group {
             Box::new(github_actions::group(format_args!(
                 "{}:{}",
                 display(&self.path),
@@ -235,11 +237,12 @@ impl<const GROUP: bool> TestStatus for PathAndRev<GROUP> {
     }
 }
 
-impl<const GROUP: bool> StatusEmitter for Gha<GROUP> {
+impl StatusEmitter for Gha {
     fn register_test(&self, path: PathBuf) -> Box<dyn TestStatus> {
-        Box::new(PathAndRev::<GROUP> {
+        Box::new(PathAndRev {
             path,
             revision: String::new(),
+            group: self.group,
         })
     }
 
@@ -252,7 +255,7 @@ impl<const GROUP: bool> StatusEmitter for Gha<GROUP> {
         // Can't aborted on gha
         _aborted: bool,
     ) -> Box<dyn Summary> {
-        struct Summarizer<const GROUP: bool> {
+        struct Summarizer {
             failures: Vec<String>,
             succeeded: usize,
             ignored: usize,
@@ -260,7 +263,7 @@ impl<const GROUP: bool> StatusEmitter for Gha<GROUP> {
             name: String,
         }
 
-        impl<const GROUP: bool> Summary for Summarizer<GROUP> {
+        impl Summary for Summarizer {
             fn test_failure(&mut self, status: &dyn TestStatus, errors: &Errors) {
                 let revision = if status.revision().is_empty() {
                     "".to_string()
@@ -274,7 +277,7 @@ impl<const GROUP: bool> StatusEmitter for Gha<GROUP> {
                     .push(format!("{}{revision}", display(status.path())));
             }
         }
-        impl<const GROUP: bool> Drop for Summarizer<GROUP> {
+        impl Drop for Summarizer {
             fn drop(&mut self) {
                 if let Some(mut file) = github_actions::summary() {
                     writeln!(file, "### {}", self.name).unwrap();
@@ -297,7 +300,7 @@ impl<const GROUP: bool> StatusEmitter for Gha<GROUP> {
             }
         }
 
-        Box::new(Summarizer::<GROUP> {
+        Box::new(Summarizer {
             failures: vec![],
             succeeded,
             ignored,