+++ tests/integrations/basic-fail/tests/actual_tests/bad_pattern_multiple.rs
@@ -1,0 +1,8 @@
+use basic_fail::add;
+
+fn main() {
+    add("42", 3);
+    //~^ ERROR: miesmätsched types
+    add("42", 3);
+    //~^ ERROR: miesmätsched types
+}